Added comment view.

+++ FalkenbergsRevyn/Controllers/CommentController.cs
@@ -12,16 +12,18 @@ namespace FalkenbergsRevyn.Controllers
         {
             _context = context;
         }
+        // GET: Comment
         public async Task<IActionResult> Index()
         {
             var comments = await _context.Comments.ToListAsync();
             return View(comments);
         }
-        public async Task<IActionResult> Detilas(int? id)
+        public async Task<IActionResult> Details(int id)
         {
             if (id == null)
             {
                 return NotFound();
+
             }
             var comment = await _context.Comments.FirstOrDefaultAsync(m => m.CommentId == id);
             if (comment == null)
@@ -30,17 +32,21 @@ namespace FalkenbergsRevyn.Controllers
             }
             return View(comment);
         }
-
+        // In case we need to add more comments
         public IActionResult Create()
         {
             return View();
         }
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("CommentId,CommentContent, Category")] Comment comment)
+        public async Task<IActionResult> Create([Bind("Content, Category")] Comment comment)
         {
             if (ModelState.IsValid)
             {
+                comment.DatePosted = DateTime.Now;
+                comment.IsAnswered = false;
+                comment.IsArchived = false;
+                comment.PostId = 1; // Need to think how to connect this to a post
                 _context.Add(comment);
                 await _context.SaveChangesAsync();
                 return RedirectToAction(nameof(Index));
@@ -67,8 +73,11 @@ namespace FalkenbergsRevyn.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var comment = await _context.Comments.FindAsync(id);
-            _context.Comments.Remove(comment);
-            await _context.SaveChangesAsync();
+            if (comment != null)
+            {
+                _context.Comments.Remove(comment);
+                await _context.SaveChangesAsync();
+            }
             return RedirectToAction(nameof(Index));
         }
     }
